Working through solving callback error
