+++ .gitignore
@@ -5,3 +5,4 @@ __pycache__/
 *.pyc
 .DS_STORE
 *.html
+.python-version
