+++ app.py
@@ -3,6 +3,7 @@ import dash
 import dash_bootstrap_components as dbc
 
 app = dash.Dash(__name__,
-                external_stylesheets=[dbc.themes.BOOTSTRAP, "https://fonts.googleapis.com/css2?family=Montserrat:wght@100;200;300;400;500;600;700&display=swap", "https://fonts.googleapis.com/css2?family=Montserrat:wght@100;200;300;400;500;600;700&family=Roboto:wght@100;300;400;500;700&display=swap"],
-                meta_tags=[{"name": "viewport", "content": "width=device-width"}],
-                suppress_callback_exceptions=True)
+  external_stylesheets=[dbc.themes.BOOTSTRAP, "https://fonts.googleapis.com/css2?family=Montserrat:wght@100;200;300;400;500;600;700&display=swap", "https://fonts.googleapis.com/css2?family=Montserrat:wght@100;200;300;400;500;600;700&family=Roboto:wght@100;300;400;500;700&display=swap"],
+  meta_tags=[{"name": "viewport", "content": "width=device-width"}]
+  , suppress_callback_exceptions=True
+)
